requirements.txt
#libraries to be installed
streamlit==1.10.0
google-generativeai

